src/main/java/com/concrete/validadorCPF/models/Cliente.java
package com.concrete.validadorCPF.models;

import org.springframework.boot.SpringApplication;
import org.springframework.boot.autoconfigure.SpringBootApplication;

public class Cliente{
    private String nome;
    private String cpf;

    public void setNome(String nome){
        this.nome = nome;
    }

    public String getNome(){
        return this.nome;
    }

    public void setCpf(String cpf){
        this.cpf = cpf;
    }

    public String getCpf(){
        return this.cpf;
    }

    public boolean validarCPF(){
        return true;
    }
}
